docker_compose/docker_postgresql/requirements.txt
requests
psycopg2-binary
configparser
